% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{bmi}
\usage{
bmi(mass, height)
}
\arguments{
\item{mass}{weight value in kg (should be between 0 and 200)}

\item{height}{height value in meter (should be between 0. and 2.5)}
}
\value{
The Body Mass Index in kg/m2
}
\description{
Function calculation the Body Mass Index
}
\examples{
bmi(mass = 70, height = 1.8)
}
