% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weight.R
\docType{data}
\name{data_weight}
\alias{data_weight}
\title{Weight and height of Star War characters}
\format{
A data frame with 87 rows and 3 variables:
\describe{
\item{name}{name of character}
\item{height}{height in meter}
\item{mass}{mass in kg}
...
}
}
\usage{
data_weight
}
\description{
A dataset #' Weight and height of Star War characters
}
\keyword{datasets}
