% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_import_excel.R
\name{multi_import_excel}
\alias{multi_import_excel}
\title{multi_import_excel Title}
\usage{
multi_import_excel(path)
}
\arguments{
\item{path}{path to the xlsx file to read}
}
\value{
list containing all the xlsx file spreadsheet tables
}
\description{
multi_import_excel Title
}
